+++ suit5/tests/templates/form_tabs.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from django.utils.translation import ugettext
+from django.utils.translation import gettext
 from suit5.tests.mixins import ModelsTestCaseMixin, UserTestCaseMixin
 from suit5.tests.models import Book, BookAdmin, test_app_label
 
@@ -15,7 +15,7 @@ app_label = test_app_label()
 
 class TabbedBookAdmin(BookAdmin):
     list_filter = ('id', 'name',)
-    suit_form_tabs = (('tab1', 'Tab1'), ('tab2', ugettext('Tab2')))
+    suit_form_tabs = (('tab1', 'Tab1'), ('tab2', gettext('Tab2')))
     suit_form_includes = None
 
 
@@ -43,7 +43,7 @@ class FormTabsTestCase(ModelsTestCaseMixin, UserTestCaseMixin):
         )
         self.get_response(self.url)
         self.assertTemplateUsed(self.response,
-                                'suit/includes/change_form_includes.html')
+                                'suit5/includes/change_form_includes.html')
         self.assertTemplateUsed(self.response, suit_form_include)
         self.assertContains(self.response,
                             '<div class="suit-include suit-tab suit-tab-tab1">')