+++ suit5/tests/templatetags/suit_tags.py
@@ -1,7 +1,6 @@
 import datetime
 from django.conf import settings
 from django.test import TestCase
-from django.utils.encoding import python_2_unicode_compatible
 from suit5 import utils
 from suit5.templatetags.suit_tags import suit_conf, suit_date, suit_time, \
     admin_url, field_contents_foreign_linked, suit_bc, suit_bc_value
@@ -10,7 +9,6 @@ from django.contrib import admin
 from django.contrib.admin.helpers import AdminReadonlyField
 
 
-@python_2_unicode_compatible
 class Country(models.Model):
     name = models.CharField(max_length=64)
 
@@ -18,7 +16,6 @@ class Country(models.Model):
         return self.name
 
 
-@python_2_unicode_compatible
 class City(models.Model):
     name = models.CharField(max_length=64)
     country = models.ForeignKey(Country, on_delete=models.CASCADE)
@@ -27,12 +24,16 @@ class City(models.Model):
         return self.name
 
 
-class CityAdmin(admin.ModelAdmin):
+class HiddenAdmin(admin.ModelAdmin):
+    def has_module_permission(self, request):
+        return False
+
+
+class CityAdmin(HiddenAdmin):
     readonly_fields = ('country',)
-    pass
 
 
-admin.site.register(Country)
+admin.site.register(Country, HiddenAdmin)
 admin.site.register(City, CityAdmin)
 
 
@@ -82,12 +83,10 @@ class SuitTagsTestCase(TestCase):
 
         # Create form
         request = None
-        form = ma.get_form(request, city)
-        form.instance = city
+        form = ma.get_form(request, city)(instance=city)
         ro_field = AdminReadonlyField(form, 'country', True, ma)
 
-        self.assertEqual(country.name,
-                         field_contents_foreign_linked(ro_field))
+        self.assertIn(country.name, field_contents_foreign_linked(ro_field))
 
         # Now it should return as link
         ro_field.model_admin.linked_readonly_fields = ('country',)