+++ suit5/tests/test.py
@@ -1,3 +1,12 @@
-from .templatetags import *
-from .templates import *
-from . import *
+from suit5.tests.config import ConfigTestCase, ConfigWithModelsTestCase  # noqa: F401
+from suit5.tests.templates.form_tabs import FormTabsTestCase  # noqa: F401
+from suit5.tests.templatetags.suit_list import SuitListTestCase  # noqa: F401
+from suit5.tests.templatetags.suit_menu import (  # noqa: F401
+    SuitMenuAdminCustomURLTestCase,
+    SuitMenuAdminI18NURLTestCase,
+    SuitMenuAdminRootURLTestCase,
+    SuitMenuTestCase,
+)
+from suit5.tests.templatetags.suit_tags import SuitTagsTestCase  # noqa: F401
+from suit5.tests.utils import UtilsTestCase  # noqa: F401
+from suit5.tests.widgets import WidgetsTestCase  # noqa: F401